+++ src/Client/ClientCore/Systems/ClientWorldEdit/ClientWorldEditInstaller.cs
@@ -27,5 +27,7 @@ public class ClientWorldEditInstaller : IWindsorInstaller
         container.Register(Component.For<ClientWorldEditState>().LifestyleSingleton());
         container.Register(Component.For<ClientWorldEditServices>().LifestyleSingleton());
         container.Register(Component.For<ClientWorldEditController>().LifestyleSingleton());
+        container.Register(Component.For<ClientWorldEditInputHandler>().LifestyleSingleton());
+        container.Register(Component.For<ClientWorldEditInternalController>().LifestyleSingleton());
     }
 }